Cast to get the format specifier right

+++ src/fsort.c
@@ -127,10 +127,10 @@ SEXP fsort(SEXP x, SEXP verboseArg) {
   t[1] = wallclock();
   double *mins = (double *)malloc(nBatch * sizeof(double));
   if (!mins)
-    error(_("Failed to allocate %d bytes in fsort() for array %s."), nBatch * sizeof(double), "mins");
+    error(_("Failed to allocate %d bytes in fsort() for array %s."), (int)(nBatch * sizeof(double)), "mins");
   double *maxs = (double *)malloc(nBatch * sizeof(double));
   if (!maxs)
-    error(_("Failed to allocate %d bytes in fsort() for array %s."), nBatch * sizeof(double), "maxs");
+    error(_("Failed to allocate %d bytes in fsort() for array %s."), (int)(nBatch * sizeof(double)), "maxs");
   const double *restrict xp = REAL(x);
   #pragma omp parallel for schedule(dynamic) num_threads(getDTthreads(nBatch, false))
   for (int batch=0; batch<nBatch; ++batch) {
